Snapshot of the previous project code
This is a commit to store a snapshot of this project. I plan on start basically from scratch with a different system architecture in mind from the start. To create this service with this method(using Kafka and maintaining different threads for each link we want to manage)  should be somewhat manageable but would cost a lot of resources, hence I am switching my approach to developing this service.